Extract common PickerCanceledError to a common path
Both `GooglePickerClient` and `OneDrivePickerClient` use
`PickerCanceledError`. @robertknight suggested to extract this error and
move it to a common location:
https://github.com/hypothesis/lms/pull/3134#discussion_r717529172
