+++ lms/static/scripts/frontend_apps/errors.js
@@ -1,0 +1,8 @@
+/**
+ * Error thrown when the user cancels file selection.
+ */
+export class PickerCanceledError extends Error {
+  constructor() {
+    super('Dialog was canceled');
+  }
+}
