+++ lms/static/scripts/frontend_apps/utils/test/onedrive-picker-client-test.js
@@ -1,5 +1,5 @@
 import { delay } from '../../../test-util/wait';
-import { PickerCanceledError } from '../google-picker-client';
+import { PickerCanceledError } from '../../errors';
 import { OneDrivePickerClient, $imports } from '../onedrive-picker-client';
 
 describe('OneDrivePickerClient', () => {
